Adds hn-components.dart
Contains Title, TimeAge, and Domain components
